+++ backend/src/main/java/com/woowacourse/gongseek/article/presentation/ArticleController.java
@@ -7,8 +7,8 @@ import com.woowacourse.gongseek.article.presentation.dto.ArticleRequest;
 import com.woowacourse.gongseek.article.presentation.dto.ArticleResponse;
 import com.woowacourse.gongseek.article.presentation.dto.ArticleUpdateRequest;
 import com.woowacourse.gongseek.article.presentation.dto.ArticleUpdateResponse;
-import com.woowacourse.gongseek.auth.presentation.AuthenticationPrinciple;
-import com.woowacourse.gongseek.auth.presentation.LoginUser;
+import com.woowacourse.gongseek.auth.presentation.anntation.AuthenticationPrinciple;
+import com.woowacourse.gongseek.auth.presentation.anntation.LoginUser;
 import com.woowacourse.gongseek.auth.presentation.dto.AppMember;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
