+++ backend/src/main/java/com/woowacourse/gongseek/auth/presentation/AuthenticationArgumentResolver.java
@@ -3,6 +3,7 @@ package com.woowacourse.gongseek.auth.presentation;
 import static org.hibernate.validator.internal.metadata.core.ConstraintHelper.PAYLOAD;
 
 import com.woowacourse.gongseek.auth.exception.HttpRequestNullException;
+import com.woowacourse.gongseek.auth.presentation.anntation.AuthenticationPrinciple;
 import com.woowacourse.gongseek.auth.presentation.dto.AppMember;
 import com.woowacourse.gongseek.auth.presentation.dto.GuestMember;
 import com.woowacourse.gongseek.auth.presentation.dto.LoginMember;
