+++ backend/src/main/java/com/woowacourse/gongseek/member/presentation/MemberController.java
@@ -1,6 +1,6 @@
 package com.woowacourse.gongseek.member.presentation;
 
-import com.woowacourse.gongseek.auth.presentation.AuthenticationPrinciple;
+import com.woowacourse.gongseek.auth.presentation.anntation.AuthenticationPrinciple;
 import com.woowacourse.gongseek.auth.presentation.dto.AppMember;
 import com.woowacourse.gongseek.member.application.MemberService;
 import com.woowacourse.gongseek.member.presentation.dto.MemberDto;
